Refactoring to support parallel shooting pattern
The parallel shooting pattern produces CT_Shot objects with different
origins.  To support that, the base class CT_ShootingPattern does not
contain anymore the origin, which a specific attribute depending on the
shooting pattern. The subclass needs to implement only getShotAt(), that
returns a CT_Shot with both origin and direction set appropriately.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>