Add proxy class to access protected members
The API was restricting the access to the global cloud. We need to access to
the global cloud in unit tests to verify for memory leak. Because the accessor
method is protected, we cannot directly call globalCloud() in a plug-in or
test, because the external class must be listed as friend. Friendship is not
transitive, meaning that we cannot list a class as friend and inherit from it
in the plug-in in order to access the method. For this reason, we use a proxy
class which is friend and provides a limited set of accessor methods.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>